pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hyperoil"
version = "2.0.0"
description = "Production-grade pair trading bot for CL/BRENTOIL on Hyperliquid DEX"
requires-python = ">=3.12"
dependencies = [
    # Async core
    "websockets>=13.0",
    "aiohttp>=3.9",
    "aiosqlite>=0.20",

    # Exchange
    "hyperliquid-python-sdk>=0.22.0",

    # Data / Quant
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.12",
    "statsmodels>=0.14",

    # Config / Validation
    "pydantic>=2.6",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Storage
    "sqlalchemy[asyncio]>=2.0",

    # Observability
    "structlog>=24.1",
    "rich>=13.7",

    # Optimization
    "optuna>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]
viz = [
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.scripts]
hyperoil = "hyperoil.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hyperoil"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
